feat: Add LinkedIn profile scraping functionality
This commit adds the functionality to scrape LinkedIn profiles of GitHub followers. The `linkedin_profiles` method is added to the `MainFollowUnfollow` class, allowing the user to specify the maximum number of LinkedIn accounts to scrape. The method retrieves the GitHub followers and then uses the `GitHubLinkedInScraper` class to scrape the LinkedIn profiles.
`unfollow` enable you to unfollow everyone with some exeptions that will be in a .jsonl.